refactoring drawerPage
