+++ aLodjinha/src/routes/DrawerScreen/styles.js
@@ -1,0 +1,19 @@
+import { StyleSheet } from 'react-native';
+
+export default StyleSheet.create({
+	menu: {
+		marginTop: 30
+	},
+	menuItem: {
+		marginLeft: 25,
+		flexDirection: 'row',
+		alignItems: 'center',
+		marginBottom: 30
+	},
+	menuText: {
+		marginLeft: 20,
+		fontWeight: 'bold',
+		fontSize: 16,
+		color: '#333'
+	}
+});
